+++ Changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Version (NEXT)
-* fixed some bugs
+* improved layout of the svg circuit visualization
+* improved HHL algorithm
 * removed duplicated code
 * increased test coverage
 * updated tutorial